:speech_balloon: 点播流程
