+++ gb28181-client/src/main/java/io/github/lunasaw/gbproxy/client/transmit/request/bye/ByeClientProcessorClient.java
@@ -1,0 +1,12 @@
+package io.github.lunasaw.gbproxy.client.transmit.request.bye;
+
+import io.github.lunasaw.sip.common.service.SipUserGenerate;
+
+/**
+ * @author luna
+ * @date 2023/11/14
+ */
+public interface ByeClientProcessorClient extends SipUserGenerate {
+
+    void closeStream(String callId);
+}
