+++ gb28181-client/src/main/java/io/github/lunasaw/gbproxy/client/transmit/request/invite/InviteRequestProcessor.java
@@ -6,12 +6,9 @@ import javax.sip.header.ContentTypeHeader;
 import javax.sip.message.Response;
 
 import gov.nist.javax.sip.message.SIPRequest;
-import io.github.lunasaw.gbproxy.client.transmit.cmd.ClientSendCmd;
 import io.github.lunasaw.sip.common.entity.*;
 import io.github.lunasaw.sip.common.enums.ContentTypeEnum;
 import io.github.lunasaw.sip.common.transmit.ResponseCmd;
-import io.github.lunasaw.sip.common.transmit.SipSender;
-import io.github.lunasaw.sip.common.utils.SipRequestUtils;
 import io.github.lunasaw.sip.common.utils.SipUtils;
 import org.springframework.stereotype.Component;
 
@@ -37,7 +34,7 @@ public class InviteRequestProcessor extends SipRequestProcessorAbstract {
     private String method = METHOD;
 
     @Resource
-    private InviteClientProcessorClient inviteClientProcessorClient;
+    private InviteProcessorClient inviteProcessorClient;
 
     /**
      * 收到Invite请求 处理
@@ -52,7 +49,7 @@ public class InviteRequestProcessor extends SipRequestProcessorAbstract {
         String userId = SipUtils.getUserIdFromToHeader(request);
 
         // 获取设备
-        FromDevice fromDevice = (FromDevice) inviteClientProcessorClient.getFromDevice();
+        FromDevice fromDevice = (FromDevice) inviteProcessorClient.getFromDevice();
 
         if (!userId.equals(fromDevice.getUserId())) {
             return;
@@ -61,8 +58,8 @@ public class InviteRequestProcessor extends SipRequestProcessorAbstract {
         String callId = SipUtils.getCallId(request);
         // 解析Sdp
         GbSessionDescription sessionDescription = (GbSessionDescription) SipUtils.parseSdp(new String(request.getRawContent()));
-        inviteClientProcessorClient.inviteSession(callId, sessionDescription);
-        String content = inviteClientProcessorClient.getAckContent(userId, sessionDescription);
+        inviteProcessorClient.inviteSession(callId, sessionDescription);
+        String content = inviteProcessorClient.getAckContent(userId, sessionDescription);
 
         String receiveIp = request.getLocalAddress().getHostAddress();
         ContentTypeHeader contentTypeHeader = ContentTypeEnum.APPLICATION_SDP.getContentTypeHeader();
