+++ gb28181-test/src/main/java/io/github/lunasaw/gbproxy/test/user/client/DefaultByeProcessorClient.java
@@ -1,0 +1,32 @@
+package io.github.lunasaw.gbproxy.test.user.client;
+
+import io.github.lunasaw.gbproxy.client.transmit.request.bye.ByeClientProcessorClient;
+import io.github.lunasaw.sip.common.entity.Device;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+/**
+ * @author luna
+ * @date 2023/11/14
+ */
+@Component
+public class DefaultByeProcessorClient implements ByeClientProcessorClient {
+
+    @Autowired
+    private FfmpegCommander ffmpegCommander;
+
+    @Override
+    public void closeStream(String callId) {
+        ffmpegCommander.closeStream(callId);
+    }
+
+    @Override
+    public Device getToDevice(String userId) {
+        return null;
+    }
+
+    @Override
+    public Device getFromDevice() {
+        return null;
+    }
+}
