+++ gb28181-test/src/test/java/io/github/lunasaw/gbproxy/test/Gb28181TestClient.java
@@ -4,7 +4,6 @@ import javax.sip.message.Request;
 
 import io.github.lunasaw.gbproxy.client.transmit.cmd.ClientSendCmd;
 import io.github.lunasaw.gbproxy.test.config.DeviceConfig;
-import io.github.lunasaw.gbproxy.test.user.client.DefaultInviteClientProcessorClient;
 import io.github.lunasaw.gbproxy.test.user.client.DefaultRegisterProcessorClient;
 import io.github.lunasaw.sip.common.utils.SipRequestUtils;
 import lombok.extern.slf4j.Slf4j;
@@ -45,7 +44,7 @@ public class Gb28181TestClient {
     public void before() {
         // 本地端口监听
         log.info("before::客户端初始化 fromDevice.ip : {} , fromDevice.port : {}", fromDevice.getIp(), fromDevice.getPort());
-        SipLayer.addListeningPoint(DeviceConfig.LOOP_IP, fromDevice.getPort());
+        SipLayer.addListeningPoint(DeviceConfig.LOOP_IP_LOCAL, fromDevice.getPort());
     }
 
     @Test
@@ -65,7 +64,7 @@ public class Gb28181TestClient {
     public void test_register_client_custom() throws Exception {
         String callId = SipRequestUtils.getNewCallId();
 
-        ToDevice instance = ToDevice.getInstance("41010500002000000001", "10.37.4.43", 8116);
+        ToDevice instance = ToDevice.getInstance("41010500002000000001", "10.37.2.198", 8116);
         instance.setPassword("bajiuwulian1006");
 
         DefaultRegisterProcessorClient.deviceMap.put("41010500002000000001", instance);
